backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
passlib[argon2]
python-jose
pydantic

